+++ NEWS.md
@@ -7,9 +7,16 @@ commas, which are a no-no in CSVs
 - Reject any rows where Cal = 1 unless Gen_QC is 11 or 12
 - Change Unique_ID to <Site>_<Date>_<Time> to make it easier to compare across sites
 
-Changes to `report_site`:
+Changes to daily stats CSV
+- Add minimum and maximum temperature
+
+Seasonal stats CSV
+- report_site now writes a seasonal stats CSV as well as including table in report
+
+Changes to report:
 - Move figure captions above figures (which looks odd!)
-- Add minimum and maximum temperature to daily stats CSV
+
+