+++ NEWS.md
@@ -1,3 +1,38 @@
+# BuzzardsBay 0.1.0.9028
+
+* Change to **MX801** import.  
+  Switch to using a YAML file (`.yml`) when importing metadata from MX801, 
+  ignoring the details sheet in the excel file that was previously
+  used.
+  
+  This change was  made to: 
+  (1) allow users to specify the deployment window in the yaml file
+  (use `calibration_start` and `calibration_end`); and 
+  (2) make the import less susceptible to possible future changes in the 
+  HOBOware details sheet.  This part of the import was complicated and 
+  likely to break with minor changes to the format.
+
+  The new `details.yml` file in the calibrated directory will have to be 
+  created by users and should have this format:
+   
+    ```
+    calibration_start: 2025-01-02 15:50:02
+    calibration_end: 2025-01-04 13:00:02
+    timezone: EST
+    serial_number: 22145899
+    ```
+  The path to an example file can be retrieved with:
+    `system.file("extdata/2025/BBC/2025-01-04/Calibrated/details.yml", package = "BuzzardsBay")`
+  
+* Change to **CSV** import. Previously the CSV export expected the 
+`calibration_start` and `calibration_end` times in the YAML file to match the
+ date and time of the first and last record in the file.
+ Now, it does not need to match and any rows before `calibration_start` 
+ or after `calibation_end` **will be dropped**.
+ 
+* Document the three import types and file formats in `qc_deployment()`/
+  
+
 # BuzzardsBay 0.1.0.9027
 
 - Fix a bug that occurred when there wasn't a gap between deployments